MiddleFork04.ssn: Middle Fork 2004 stream temperature dataset

The MiddleFork04.ssn data folder contains the spatial, attribute, and topological information needed to construct an SSN object using the SSN2 package.

Source

edges are a modified version of the United States National Hydrography Dataset (http://nhd.usgs.gov/). sites, pred1km and CapeHorn are unpublished United States Forest Service data.

Details

The MiddleFork04.ssn folder contains five geopackages:

edges: geopackage with LINESTRING geometry representing the stream network

sites: geopackage with POINT geometry representing the observed site locations

pred1km: geopackage with POINT geometry prediction site locations at approximately 1km intervals throughout the stream network

CapeHorn: geopackage with POINT geometry representing prediction site locations on the Cape Horn River

The MiddleFork04.ssn includes two text files, netID1.txt and netID2.txt, which contain the topological information for the two stream networks in the Middle Fork 2004 dataset.

The distance folder contains four folders that store the hydrologic distance matrices for each of the point datasets (obs, CapeHorn, and pred1km). See ssn_create_distmat() for a detailed description of the distance matrix file structure.

Attribute data is also stored within each of the spatial datasets. The column names are defined as follows:

edges:

rid: Reach identifier

COMID: Common identifier of an NHD feature or relationship

GNIS_Name: Feature name as found in the Geographic Names Information System

REACHCODE: Unique identifier for a reach. The first 8 digits contain the identifier for the HUC8 and the last 6 digits are a unique within-HUC8 identifier for the reach

FTYPE: three-digit integer used to classify hydrography features in the NHD and define subtypes

FCODE: Numeric code that contains the feature type and its attributes as found in the NHDFCode lookup table

AREAWTMAP: Area weighted mean annual precipitation (mm) at the lowermost location on the edge

SLOPE: Slope of the edge (cm/cm)

rcaAreaKm2: Reach contributing area (km2) for each edge feature. The RCA represents the land area that drains directly to the edge feature

h2oAreaKm2: Watershed area (km2) for the lowermost location on the edge feature

areaPI: Segment proportional influence value, calculated using watershed area (h2oAreaKm2)

afvArea: Additive function value, calculated using areaPI

upDist: Distance from the stream outlet (most downstream location in the the stream network) to the uppermost location on the line segment

Length: Length of line segment (m)

netID: Network identifier

sites:

rid: Reach identifier of the edge the site resides on

pid: Point identifier

STREAMNAME: Stream name

COMID: Common identifier of an NHD feature or relationship

AREAWTMAP: Area weighted mean annual precipitation (mm) at lowermost location on the line segment where the site resides

SLOPE: Slope of the line segment (cm/cm) where the site resides

ELEV_DEM: Elevation at the site based on a 30m DEM

Source: Source of the data - relates to the ID field of the source table

Summer_mn: Overall summer mean temperature (C) of the deployment

MaxOver20: Binary variable: 1 represents the maximum summer temperature was greater than 20C and 0 indicates that it was less than 20C

C16: Number of times daily stream temperature exceeded 16C

C20: Number of times daily stream temperature exceeded 20C

C24: Number of times daily stream temperature exceeded 24C

FlowCMS: Average stream flow (cubic meters per sec) for August, by year, from 1950-2010 across 9 USGS gauges in the region

AirMEANc: Average mean air temperature (C) from July 15 - August 31, from 1980-2009 across 10 COOP air stations within the domain

AirMWMTc: Average maximum air temperature (C) from July 15 - August 31, from 1980-2009 across 10 COOP air stations within the domain. MWMT = maximum 7-day moving average of the maximum daily temperature (i.e. maximum of all the 7-day maximums)

rcaAreaKm2: Reach contributing area (km2) for each edge feature the site resides on. The RCA represents the land area that drains directly to the edge feature

h2oAreaKm2: Watershed area (km2) for the lowermost location on the edge feature the site resides on

ratio: Site ratio value; provides the proportional distance from the downstream node of the edge to the site location

upDist: Distance upstream from the stream outlet (m)

afvArea: Additive function value calculated using watershed area (h2oAreaKm2)

locID: Location identifier

netID: Stream network identifier

snapdist: Distance (m) that site was moved when it was snapped to the edges

pred1km and CapeHorn:

rid: Reach identifier of the edge the site resides on

pid: Point identifier

COMID: Common identifier of an NHD feature or relationship

GNIS_Name: Feature name of the edge the site resides on, as found in the Geographic Names Information System

AREAWTMAP: Area weighted mean annual precipitation (mm) at lowermost location on the line segment where the site resides

SLOPE: Slope of the line segment (cm/cm) where the site resides

ELEV_DEM: Elevation at the site based on a 30m DEM

rcaAreaKm2: Reach contributing area (km2) for each edge feature the site resides on. The RCA represents the land area that drains directly to the edge feature

h2oAreaKm2: Watershed area (km2) for the lowermost location on the edge feature the site resides on

ratio: Site ratio value; provides the proportional distance along the edge to the site location

upDist: Distance upstream from the stream outlet (m)

afvArea: Additive function value calculated using watershed area (h2oAreaKm2)

locID: Location identifier

netID: Stream network identifier

snapdist: Distance (m) that site was moved when it was snapped to the edges

FlowCMS: Average stream flow (cubic meters per sec) for August, by year, from 1950-2010 across 9 USGS gauges in the region

AirMEANc: Average mean air temperature (C) from July 15 - August 31, from 1980-2009 across 10 COOP air stations within the domain

AirMWMTc: Average maximum air temperature (C) from July 15 - August 31, from 1980-2009 across 10 COOP air stations within the domain. MWMT = maximum 7-day moving average of the maximum daily temperature(i.e. maximum of all the 7-day maximums)

See also