Extract fitted model coefficients

coef extracts fitted model coefficients from fitted model objects. coefficients is an alias for it.

Usage

# S3 method for class 'ssn_lm'
coef(object, type = "fixed", ...)

# S3 method for class 'ssn_lm'
coefficients(object, type = "fixed", ...)

# S3 method for class 'ssn_glm'
coef(object, type = "fixed", ...)

# S3 method for class 'ssn_glm'
coefficients(object, type = "fixed", ...)

Arguments

object

A fitted model object from ssn_lm() or ssn_glm().

type

"fixed" for fixed effect coefficients, "tailup" for tailup covariance parameter coefficients, "taildown" for taildown covariance parameter coefficients, "euclid" for Euclidean covariance parameter coefficients, "nugget" for nugget covariance parameter coefficients, "dispersion" for the dispersion parameter coefficient (ssn_glm() objects), "randcov" for random effect variance coefficients, or "ssn" for all of the tailup, taildown, Euclidean, nugget, and dispersion (ssn_glm() objects) parameter coefficients. Defaults to "fixed".

...

Other arguments. Not used (needed for generic consistency).

Value

A named vector of coefficients.

Examples

# Copy the mf04p .ssn data to a local directory and read it into R
# When modeling with your .ssn object, you will load it using the relevant
# path to the .ssn data on your machine
copy_lsn_to_temp()
temp_path <- paste0(tempdir(), "/MiddleFork04.ssn")
mf04p <- ssn_import(temp_path, overwrite = TRUE)

ssn_mod <- ssn_lm(
  formula = Summer_mn ~ ELEV_DEM,
  ssn.object = mf04p,
  tailup_type = "exponential",
  additive = "afvArea"
)
coef(ssn_mod)
#> (Intercept)    ELEV_DEM 
#>  80.8578372  -0.0341245 
coef(ssn_mod, type = "tailup")
#>           de        range 
#> 1.390296e+00 1.306032e+05 
#> attr(,"class")
#> [1] "tailup_exponential"
coefficients(ssn_mod)
#> (Intercept)    ELEV_DEM 
#>  80.8578372  -0.0341245