Glance at a fitted model object

Returns a row of model summaries from a fitted model object. Glance returns the same number of columns for all models and estimation methods.

Usage

# S3 method for class 'ssn_lm'
glance(x, ...)

# S3 method for class 'ssn_glm'
glance(x, ...)

Arguments

x

A fitted model object from ssn_lm() or ssn_glm().

...

Other arguments. Not used (needed for generic consistency).

Value

A single-row tibble with columns

n The sample size.

p The number of fixed effects.

npar The number of estimated covariance parameters.

value The optimized value of the fitting function

AIC The AIC.

AICc The AICc.

logLik The log-likelihood

deviance The deviance.

pseudo.r.squared The pseudo r-squared

Examples

# Copy the mf04p .ssn data to a local directory and read it into R
# When modeling with your .ssn object, you will load it using the relevant
# path to the .ssn data on your machine
copy_lsn_to_temp()
temp_path <- paste0(tempdir(), "/MiddleFork04.ssn")
mf04p <- ssn_import(temp_path, overwrite = TRUE)

ssn_mod <- ssn_lm(
  formula = Summer_mn ~ ELEV_DEM,
  ssn.object = mf04p,
  tailup_type = "exponential",
  additive = "afvArea"
)
glance(ssn_mod)
#> # A tibble: 1 × 9
#>       n     p  npar value   AIC  AICc logLik deviance pseudo.r.squared
#>   <int> <dbl> <int> <dbl> <dbl> <dbl>  <dbl>    <dbl>            <dbl>
#> 1    45     2     3  76.9  82.9  83.5  -38.4     43.0            0.585