Compute leverage (hat) values

Compute the leverage (hat) value for each observation from a fitted model object.

Usage

# S3 method for ssn_lm
hatvalues(model, ...)

# S3 method for ssn_glm
hatvalues(model, ...)

Arguments

model

A fitted model object from ssn_lm() or ssn_glm().

...

Other arguments. Not used (needed for generic consistency).

Value

A vector of leverage (hat) values for each observation from the fitted model object.

Details

Leverage values measure how far an observation's explanatory variables are relative to the average of the explanatory variables. In other words, observations with high leverage are typically considered to have an extreme or unusual combination of explanatory variables. Leverage values are the diagonal of the hat (projection) matrix. The larger the hat value, the larger the leverage.

See also

Examples

# Copy the mf04p .ssn data to a local directory and read it into R
# When modeling with your .ssn object, you will load it using the relevant
# path to the .ssn data on your machine
copy_lsn_to_temp()
temp_path <- paste0(tempdir(), "/MiddleFork04.ssn")
mf04p <- ssn_import(temp_path, overwrite = TRUE)

ssn_mod <- ssn_lm(
  formula = Summer_mn ~ ELEV_DEM,
  ssn.object = mf04p,
  tailup_type = "exponential",
  additive = "afvArea"
)
hatvalues(ssn_mod)
#>            1            2            3            4            5            6 
#> 1.565733e-01 9.237419e-02 3.014500e-02 7.981493e-03 1.046372e-02 3.495971e-03 
#>            7            8            9           10           11           12 
#> 1.793629e-02 1.693222e-02 8.311967e-03 5.612972e-02 1.031498e-01 6.860116e-02 
#>           13           14           15           16           17           18 
#> 1.109540e-02 4.019243e-02 1.253507e-01 9.102368e-02 7.023914e-02 2.202866e-02 
#>           19           20           21           22           23           24 
#> 4.509632e-03 2.293475e-03 3.162038e-03 2.476739e-03 1.356383e-02 5.100491e-02 
#>           25           26           27           28           29           30 
#> 1.395086e-02 2.156846e-02 9.548471e-02 1.553130e-01 8.095946e-03 3.568117e-02 
#>           31           32           33           34           35           36 
#> 3.536055e-02 1.067164e-01 4.896040e-02 2.389906e-02 7.092277e-05 1.122144e-02 
#>           37           38           39           40           41           42 
#> 2.463285e-02 1.287318e-01 8.437105e-02 6.881851e-02 1.810541e-02 1.149500e-02 
#>           43           44           45 
#> 9.801228e-03 7.212308e-02 1.656265e-02