MiddleFork04.ssn: Middle Fork 2004 stream temperature dataset

The MiddleFork04.ssn data folder contains the spatial, attribute, and topological information needed to construct an SSN object using the SSN2 package.

Source

edges are a modified version of the United States National Hydrography Dataset (http://nhd.usgs.gov/). sites, pred1km, CapeHorn and Knapp are unpublished United States Forest Service data.

Details

The MiddleFork04.ssn folder contains five shapefiles:

edges: polyline shapefile representing the stream network

sites: point shapefile representing the observed site locations

pred1km: point shapefile representing prediction site locations at approximately 1km intervals throughout the stream network

Knapp: point shapefile representing prediction site locations on the Knapp River

CapeHorn: point shapefile representing prediction site locations on the Cape Horn River

The MiddleFork04.ssn includes one text file, netID1.txt, which contains the topological information for the stream network in the Middle Fork 2004 dataset.

The distance folder contains four folders that store the hydrologic distance matrices for each of the point shapefiles (obs, CapeHorn, Knapp, and pred1km). See ssn_create_distmat() for a detailed description of the distance matrix file structure.

Attribute data is also stored within each of the spatial datasets. The column names are defined as follows:

edges:

COMID: Common identifier of an NHD feature or relationship

GNIS_Name: Feature name as found in the Geographic Names Information System

REACHCODE: Unique identifier for a reach. The first 8 digits contain the identfier for the HUC8 and the last 6 digits are a unique within-HUC8 identifier for the reach

FTYPE: three-digit integer used to classify hydrography features in the NHD and define subtypes

FCODE: Numeric code that contains the feature type and its attributes as found in the NHDFCode lookup table

CDRAINAG: Cumulative drainage area (km2) for the lowermost location on the edge

AREAWTMAP: Area weighted mean annual precipitation (mm) at the lowermost location on the edge

SLOPE: Slope of the edge (cm/cm)

h2oAreaKm2: Watershed area (km2) for the lowermost location on the line segment

rid: Reach identifier

areaPI: Segment proportional influence value, calculated using watershed area (h2oAreaKm2)

afvArea: Additive function value, calculated using areaPI

upDist: Distance from the stream outlet (most downstream location in the the stream network) to the uppermost location on the line segment

Length: Length of line segment (m)

netID: Network identifier

sites:

STREAMNAME: Stream name

COMID: Common identifier of an NHD feature or relationship

CDRAINAG: Cumulative drainage area (km2)

AREAWTMAP: Area weighted mean annual precipitation (mm) at lowermost location on the line segment where the site resides

SLOPE: Slope of the line segment (cm/cm) where the site resides

ELEV_DEM: Elevation at the site based on a 30m DEM

Source: Source of the data - relates to the ID field of the source table

Summer_mn: Overall summer mean termperature (C) of the deployment

MaxOver20: Binary variable: 1 represents the maximum summer temperature was greater than 20C and 0 indicates that it was less than 20C

C16: Number of times daily stream temperature exceeded 16C

C20: Number of times daily stream temperature exceeded 20C

C24: Number of times daily stream temperature exceeded 24C

FlowCMS: Average stream flow (cubic meters per sec) for August, by year, from 1950-2010 across 9 USGS gauges in the region

AirMEANc: Average mean air temperature (C) from July 15 - August 31, from 1980-2009 across 10 COOP air stations within the domain

AirMWMTc: Average maximum air temperature (C) from July 15 - August 31, from 1980-2009 across 10 COOP air stations within the domain. MWMT = maximum 7-day moving average of the maximum daily temperature (i.e. maximum of all the 7-day maximums)

NEAR_X: x coordinate

NEAR_Y: y coordinate

rid: Reach identifier of the edge the site resides on

ratio: Site ratio value; provides the proportional distance along the edge to the site location

upDist: Distance upstream from the stream outlet (m)

afvArea: Additive function value calculated using waterhsed area (h2oAreaKm2)

locID: Location identifier

netID: Stream network identifier

pid: Point identifier

pred1km, CapeHorn, and Knapp:

COMID: Common identifier of an NHD feature or relationship

GNIS_Name: Feature name of the edge the site resides on, as found in the Geographic Names Information System

CDRAINAG: Cumulative drainage area (km2)

AREAWTMAP: Area weighted mean annual precipitation (mm) at lowermost location on the line segment where the site resides

SLOPE: Slope of the line segment (cm/cm) where the site resides

ELEV_DEM: Elevation at the site based on a 30m DEM

NEAR_X: x coordinate

NEAR_Y: y coordinate

rid: Reach identifier of the edge the site resides on

ratio: Site ratio value; provides the proportional distance along the edge to the site location

upDist: Distance upstream from the stream outlet (m)

afvArea: Additive function value calculated using watershed area (h2oAreaKm2)

locID: Location identifier

netID: Stream network identifier

pid: Point identifier

FlowCMS: Average stream flow (cubic meters per sec) for August, by year, from 1950-2010 across 9 USGS gauges in the region

AirMEANc: Average mean air temperature (C) from July 15 - August 31, from 1980-2009 across 10 COOP air stations within the domain

AirMWMTc: Average maximum air temperature (C) from July 15 - August 31, from 1980-2009 across 10 COOP air stations within the domain. MWMT = maximum 7-day moving average of the maximum daily temperature(i.e. maximum of all the 7-day maximums)

See also