Extract the model frame from a fitted model object

Extract the model frame from a fitted model object.

Usage

# S3 method for ssn_lm
model.frame(formula, ...)

# S3 method for ssn_glm
model.frame(formula, ...)

Arguments

formula

A fitted model object from ssn_lm() or ssn_glm().

...

Other arguments. Not used (needed for generic consistency).

Value

A model frame that contains the variables used by the formula for the fitted model object.

See also

Examples

# Copy the mf04p .ssn data to a local directory and read it into R
# When modeling with your .ssn object, you will load it using the relevant
# path to the .ssn data on your machine
copy_lsn_to_temp()
temp_path <- paste0(tempdir(), "/MiddleFork04.ssn")
mf04p <- ssn_import(temp_path, overwrite = TRUE)

ssn_mod <- ssn_lm(
  formula = Summer_mn ~ ELEV_DEM,
  ssn.object = mf04p,
  tailup_type = "exponential",
  additive = "afvArea"
)
model.frame(ssn_mod)
#>    Summer_mn ELEV_DEM
#> 1      11.38     1977
#> 2      10.72     1984
#> 3      10.43     1993
#> 4      10.13     2007
#> 5      10.12     2009
#> 6       9.81     2012
#> 7       9.76     2023
#> 8       9.77     2023
#> 9       9.53     2026
#> 10     12.57     1988
#> 11     12.42     2013
#> 12     14.10     2015
#> 13     13.22     2038
#> 14     13.08     2052
#> 15     12.19     2066
#> 16     11.32     2085
#> 17     12.11     1975
#> 18     11.82     1982
#> 19     11.91     1999
#> 20     11.41     2002
#> 21     11.71     1997
#> 22     11.54     2006
#> 23     11.14     2026
#> 24     12.06     2036
#> 25     12.09     2046
#> 26     11.96     2050
#> 27     11.02     2071
#> 28     11.16     2016
#> 29      8.75     2034
#> 30      9.02     2042
#> 31     14.61     1958
#> 32     14.92     1947
#> 33     14.72     1952
#> 34     15.22     1923
#> 35     14.49     1932
#> 36     15.29     1940
#> 37     15.07     1940
#> 38     14.94     1945
#> 39     14.99     1948
#> 40     15.02     1950
#> 41     15.11     1949
#> 42     14.70     1950
#> 43     14.58     1951
#> 44     13.70     2015
#> 45     10.21     2077